Created include/klee/Core directory and moved appropriate files direc\ tly in lib/Core
