+++ lib/Core/AddressSpace.h
@@ -11,6 +11,7 @@
 #define KLEE_ADDRESSSPACE_H
 
 #include "Memory.h"
+
 #include "klee/Expr/Expr.h"
 #include "klee/ADT/ImmutableMap.h"
 #include "klee/System/Time.h"
