+++ lib/Core/MergeHandler.cpp
@@ -7,12 +7,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "klee/MergeHandler.h"
+#include "MergeHandler.h"
 
 #include "CoreStats.h"
+#include "ExecutionState.h"
 #include "Executor.h"
 #include "Searcher.h"
-#include "klee/ExecutionState.h"
 
 namespace klee {
