+++ lib/Core/MergeHandler.h
