+++ lib/Core/PTree.cpp
@@ -9,7 +9,8 @@
 
 #include "PTree.h"
 
-#include "klee/ExecutionState.h"
+#include "ExecutionState.h"
+
 #include "klee/Expr/Expr.h"
 #include "klee/Expr/ExprPPrinter.h"
