+++ lib/Core/Searcher.cpp
@@ -10,21 +10,23 @@
 #include "Searcher.h"
 
 #include "CoreStats.h"
+#include "ExecutionState.h"
 #include "Executor.h"
+#include "MergeHandler.h"
 #include "PTree.h"
 #include "StatsTracker.h"
 
-#include "klee/ExecutionState.h"
-#include "klee/MergeHandler.h"
-#include "klee/Statistics.h"
+#include "klee/ADT/DiscretePDF.h"
+#include "klee/ADT/RNG.h"
+#include "klee/Statistics/Statistics.h"
 #include "klee/Module/InstructionInfoTable.h"
 #include "klee/Module/KInstruction.h"
 #include "klee/Module/KModule.h"
-#include "klee/ADT/DiscretePDF.h"
-#include "klee/ADT/RNG.h"
+#include "klee/Statistics.h"
+#include "klee/Support/ErrorHandling.h"
 #include "klee/Support/ModuleUtil.h"
 #include "klee/System/Time.h"
-#include "klee/Support/ErrorHandling.h"
+
 #include "llvm/IR/CallSite.h"
 #include "llvm/IR/Constants.h"
 #include "llvm/IR/Instructions.h"
